Split listings in ch9
