--- listings/ch09/listing_11.js
@@ -1,3 +1,0 @@
-export function hasOwnProperty(obj, prop) {
-  return Object.prototype.hasOwnProperty.call(obj, prop)
-}
