--- listings/ch09/listing_19.js
@@ -1,29 +1,0 @@
-function patchElement(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/) { // --1--
-  const el = oldVdom.el
-  const {
-    class: oldClass,
-    style: oldStyle,
-    on: oldEvents,
-    ...oldAttrs
-  } = oldVdom.props
-  const {
-    class: newClass,
-    style: newStyle,
-    on: newEvents,
-    ...newAttrs
-  } = newVdom.props
-  const { listeners: oldListeners } = oldVdom
-
-  patchAttrs(el, oldAttrs, newAttrs)
-  patchClasses(el, oldClass, newClass)
-  patchStyles(el, oldStyle, newStyle)
-  newVdom.listeners = patchEvents(
-    el,
-    oldListeners,
-    oldEvents,
-    newEvents,
-    // --add--
-    hostComponent // --2--
-    // --add--
-  )
-}
