+++ listings/ch10/listing_04.js
@@ -1,17 +1,22 @@
-// --add--
-import { extractComponentProps } from './utils/props'
-// --add--
+export function mountDOM(vdom, parentEl, index/*--add--*/, hostComponent = null/*--add--*/) { // --1--
+  switch (vdom.type) {
+    case DOM_TYPES.TEXT: {
+      createTextNode(vdom, parentEl, index)
+      break
+    }
 
-// --snip-- //
+    case DOM_TYPES.ELEMENT: {
+      createElementNode(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/) // --2--
+      break
+    }
 
-// --add--
-function createComponentNode(vdom, parentEl, index, hostComponent) {
-  const Component = vdom.tag
-  const props = extractComponentProps(vdom)
-  const component = new Component(props)
+    case DOM_TYPES.FRAGMENT: {
+      createFragmentNodes(vdom, parentEl, index/*--add--*/, hostComponent/*--add--*/) // --3--
+      break
+    }
 
-  component.mount(parentEl, index)
-  vdom.component = component
-  vdom.el = component.firstElement
+    default: {
+      throw new Error(`Can't mount DOM of type: ${vdom.type}`)
+    }
+  }
 }
-// --add--
