+++ listings/ch10/listing_09.js
@@ -1,26 +1,29 @@
-export function defineComponent({ render, state, ...methods }) {
-  class Component {
-    #isMounted = false
-    #vdom = null
-    #hostEl = null
+function patchElement(oldVdom, newVdom/*--add--*/, hostComponent/*--add--*/) { // --1--
+  const el = oldVdom.el
+  const {
+    class: oldClass,
+    style: oldStyle,
+    on: oldEvents,
+    ...oldAttrs
+  } = oldVdom.props
+  const {
+    class: newClass,
+    style: newStyle,
+    on: newEvents,
+    ...newAttrs
+  } = newVdom.props
+  const { listeners: oldListeners } = oldVdom
+
+  patchAttrs(el, oldAttrs, newAttrs)
+  patchClasses(el, oldClass, newClass)
+  patchStyles(el, oldStyle, newStyle)
+  newVdom.listeners = patchEvents(
+    el,
+    oldListeners,
+    oldEvents,
+    newEvents,
     // --add--
-    #eventHandlers = null
-    #parentComponent = null
+    hostComponent // --2--
     // --add--
-
-    constructor(props = {}/*--add--*/, eventHandlers = {}, parentComponent = null/*--add--*/) {
-      this.props = props
-      this.state = state ? state(props) : {}
-      // --add--
-      this.#eventHandlers = eventHandlers
-      this.#parentComponent = parentComponent
-      // --add--
-    }
-
-    // --snip-- //
-  }
-
-  // --snip-- //
-
-  return Component
+  )
 }
