+++ listings/ch10/listing_11.js
@@ -1,15 +1,23 @@
-function createComponentNode(vdom, parentEl, index, hostComponent) {
-  const Component = vdom.tag
-  // --remove--
-  const props = extractComponentProps(vdom)
-  const component = new Component(props)
-  // --remove--
-  // --add--
-  const { props, events } = extractComponentProps(vdom)
-  const component = new Component(props, events, hostComponent)
-  // --add--
+export function defineComponent({ render, state, ...methods }) {
+  class Component {
+    // --snip-- //
 
-  component.mount(parentEl, index)
-  vdom.component = component
-  vdom.el = component.firstElement
+    mount(hostEl, index = null) {
+      if (this.#isMounted) {
+        throw new Error('Component is already mounted')
+      }
+      
+      this.#vdom = this.render()
+      mountDOM(this.#vdom, hostEl, index/*--add--*/, this/*--add--*/) // --1--
+      
+      this.#hostEl = hostEl
+      this.#isMounted = true
+    }
+    
+    // --snip-- //
+  }
+
+  // --snip-- //
+
+  return Component
 }
