+++ listings/ch10/listing_12.js
@@ -1,49 +1,60 @@
+import { destroyDOM } from './destroy-dom'
+import { DOM_TYPES, extractChildren } from './h'
+import { mountDOM } from './mount-dom'
+import { patchDOM } from './patch-dom'
+import { hasOwnProperty } from './utils/objects'
+
 export function defineComponent({ render, state, ...methods }) {
   class Component {
     #isMounted = false
     #vdom = null
     #hostEl = null
-    #eventHandlers = null
-    #parentComponent = null
-    // --add--
-    #dispatcher = new Dispatcher()
-    #subscriptions = []
-    // --add--
 
-    // --snip-- //
+    constructor(props = {}) {
+      this.props = props
+      this.state = state ? state(props) : {}
+    }
+
+    get elements() {
+      if (this.#vdom == null) {
+        return []
+      }
+
+      if (this.#vdom.type === DOM_TYPES.FRAGMENT) {
+        return extractChildren(this.#vdom).map((child) => child.el)
+      }
+
+      return [this.#vdom.el]
+    }
+
+    get firstElement() {
+      return this.elements[0]
+    }
+
+    get offset() {
+      return Array.from(this.#hostEl.children).indexOf(this.firstElement)
+    }
+    
+    updateState(state) {
+      this.state = { ...this.state, ...state }
+      this.#patch()
+    }
+
+    render() {
+      return render.call(this)
+    }
 
     mount(hostEl, index = null) {
       if (this.#isMounted) {
         throw new Error('Component is already mounted')
       }
-
+      
       this.#vdom = this.render()
       mountDOM(this.#vdom, hostEl, index, this)
-      // --add--
-      this.#wireEventHandlers()
-      // --add--
-
-      this.#isMounted = true
+      
       this.#hostEl = hostEl
+      this.#isMounted = true
     }
-
-    // --add--
-    #wireEventHandlers() {
-      this.#subscriptions = Object.entries(this.#eventHandlers).map(
-        ([eventName, handler]) => this.#wireEventHandler(eventName, handler)
-      )
-    }
-
-    #wireEventHandler(eventName, handler) {
-      return this.#dispatcher.subscribe(eventName, (payload) => {
-        if (this.#parentComponent) {
-          handler.call(this.#parentComponent, payload)
-        } else {
-          handler(payload)
-        }
-      })
-    }
-    // --add--
     
     unmount() {
       if (!this.#isMounted) {
@@ -51,23 +62,31 @@ export function defineComponent({ render, state, ...methods }) {
       }
 
       destroyDOM(this.#vdom)
-      // --add--
-      this.#subscriptions.forEach((unsubscribe) => unsubscribe())
-      // --add--
 
       this.#vdom = null
-      this.#isMounted = false
       this.#hostEl = null
-      // --add--
-      this.#subscriptions = []
-      // --add--
+      this.#isMounted = false
     }
 
-    // --snip-- //
+    #patch() {
+      if (!this.#isMounted) {
+        throw new Error('Component is not mounted')
+      }
+
+      const vdom = this.render()
+      this.#vdom = patchDOM(this.#vdom, vdom, this.#hostEl, this)
+    }
   }
 
-  // --snip-- //
+  for (const methodName in methods) {
+    if (hasOwnProperty(Component, methodName)) {
+      throw new Error(
+        `Method "${methodName}()" already exists in the component. Can't override existing methods.`
+      )
+    }
+
+    Component.prototype[methodName] = methods[methodName]
+  }
 
   return Component
 }
-
