+++ listings/ch11/listing_01.js
@@ -1,8 +1,27 @@
-import {
-  createApp,
+export const DOM_TYPES = {
+  TEXT: 'text',
+  ELEMENT: 'element',
+  FRAGMENT: 'fragment',
   // --add--
-  defineComponent,
+  COMPONENT: 'component',
   // --add--
-  h,
-  hFragment,
-} from 'https://unpkg.com/fe-fwk@/*--add--*/3/*--add--*/'
+}
+
+export function h(tag, props = {}, children = []) {
+  // --add--
+  const type =
+    typeof tag === 'string' ? DOM_TYPES.ELEMENT : DOM_TYPES.COMPONENT
+  // --add--
+
+  return {
+    tag,
+    props,
+    // --add--
+    type,
+    // --add--
+    children: mapTextNodes(withoutNulls(children)),
+    // --remove--
+    type: DOM_TYPES.ELEMENT,
+    // --remove--
+  }
+}
