+++ listings/ch11/listing_08.js
@@ -1,1 +1,7 @@
-createApp(App).mount(document.body)
+function patchComponent(oldVdom, newVdom) {
+  const { component } = oldVdom
+  const { props: newProps } = newVdom
+
+  newVdom.component = component
+  component.updateProps(newProps)
+}
