+++ listings/ch12/listing_08.js
@@ -1,0 +1,1 @@
+createApp(App).mount(document.body)
